Adjust code design
